feat: improve choice of radix for Pippenger's algorithm (#85)
* add blob_array

* fill in blob array

* rework multiexponentiation code

* rework multiexponentiation code

* add test coverage for span ops

* fix typo

* rework extract_digit

* fix extract digit function

* rework multiproduct reduction

* rework multiexponentiation

* rework multiproduct functions

* rework digit code

* fix is_digit_zero

* fix zero digit test

* rework multiexponentiation code

* rework multiproduct table

* drop dead code

* support iterating over blobs

* rework multiproduct table code

* rework pippenger code

* add span iterator

* add iterator support to blob_array

* rework multiexponentiation code

* rework multiexponentiation

* rework multiexponentiation

* rework multiexponentiation

* rework multiexponentiation code

* rework multiexponentiation code

* rework multiexponentiation tests

* rework multiexponentiation tests

* rework multiexponentiation tests

* rework multiexponentiation tests

* rework multiexponentiation tests

* rework multiexponentiation code

* drop dead code

* rename

* refactor multiexponentiation code

* rework multiexponentiation code

* refactor multiexp code

* add test coverage

* fill in testing

* rework multiexponentiation

* add test_bit

* refactor multiexponentiation code

* rework multiexponentiation

* drop dead code

* drop dead code

* rework multiexponentiation

* fix format

* drop dead code

* drop dead code

* drop dead code

* drop dead code

* update benchmark results

* update radix testing

+++ sxt/base/iterator/counting_iterator.h
@@ -25,7 +25,9 @@ public:
 
   void advance(ptrdiff_t delta) noexcept { counter_ += delta; }
 
-  int distance_to(counting_iterator other) const noexcept { return other.counter_ - counter_; }
+  ptrdiff_t distance_to(counting_iterator other) const noexcept {
+    return other.counter_ - counter_;
+  }
 
 private:
   T counter_{0};
